refactor(components): relocate palette components to dedicated folder
- Move CommandPalette to Palletes/CommandPalette
- Move FileFinder to Palletes/FileFinderPallete
- Update import paths across affected components
- Fix relative imports after relocation
